PDBList update method should cope with missing or un-moveable obsolete files (see Bug 2945)

+++ Bio/PDB/PDBList.py
@@ -293,7 +293,14 @@ class PDBList:
             else:
                 old_file = self.local_pdb + os.sep + pdb_code[1:3] + os.sep + 'pdb%s.ent'%(pdb_code)
                 new_file = self.obsolete_pdb + os.sep + pdb_code[1:3] + os.sep + 'pdb%s.ent'%(pdb_code)
-            shutil.move(old_file, new_file)
+            #If the old file doesn't exist, maybe someone else moved it
+            #or deleted it already. Should we issue a warning?
+            if os.path.isfile(old_file) :
+                try :
+                    shutil.move(old_file, new_file)
+                except :
+                    warnings.warn("Could not move %s to obsolete folder" \
+                                  % pdb_code, RuntimeWarning)
 
 
     def download_entire_pdb(self,listfile=None):
